# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PAES Tutor API
  description: |
    API REST para la plataforma de preparación PAES Tutor.
    
    Esta API proporciona endpoints para:
    - Gestión de exámenes y preguntas
    - Seguimiento de intentos y progreso
    - Materiales de estudio
    - Recomendaciones personalizadas
    - Búsqueda global
    - Análisis y métricas
    
    ## Autenticación
    
    La API utiliza NextAuth.js para autenticación. Los endpoints protegidos requieren
    una sesión válida mediante cookies HTTP-only.
    
    ## Rate Limiting
    
    Los endpoints están protegidos con rate limiting usando Upstash Redis:
    - Endpoints públicos: 100 requests/minuto
    - Endpoints autenticados: 200 requests/minuto
    - Endpoints de escritura: 50 requests/minuto
    
    ## Validación
    
    Todas las solicitudes y respuestas son validadas usando Zod schemas para garantizar
    type safety y seguridad.
  version: 1.0.0
  contact:
    name: PAES Tutor Support
    email: support@paestutor.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Servidor de desarrollo local
  - url: https://paestutor.com/api
    description: Servidor de producción
...
paths:
...
  /webhooks:
    get:
      tags:
        - Notas
      summary: Listar webhooks del estudiante
      description: |
        Retorna la lista de webhooks configurados por el estudiante.
        Incluye estadísticas de deliveries recientes.
      operationId: listWebhooks
      security:
        - cookieAuth: []
      parameters:
        - name: noteId
          in: query
          description: Filtrar webhooks por nota específica
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Lista de webhooks
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhooks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Webhook'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Notas
      summary: Crear un nuevo webhook
      description: |
        Crea un nuevo webhook para recibir notificaciones de eventos de versiones.
        Requiere URL válida y al menos un evento a escuchar.
      operationId: createWebhook
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
                - events
              properties:
                url:
                  type: string
                  format: uri
                  maxLength: 2048
                  description: URL del webhook (HTTPS en producción)
                secret:
                  type: string
                  maxLength: 256
                  description: Secreto opcional para firmar payloads (HMAC)
                events:
                  type: array
                  items:
                    type: string
                    enum: ['version.created', 'version.restored', 'version.updated', 'version.deleted', 'version.marked_important', 'version.named', '*']
                  minItems: 1
                  maxItems: 20
                  description: Eventos a escuchar ('*' para todos)
                noteId:
                  type: string
                  description: ID de nota específica (opcional, null para global)
                description:
                  type: string
                  maxLength: 500
                  description: Descripción del webhook
      responses:
        '201':
          description: Webhook creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhook:
                    $ref: '#/components/schemas/Webhook'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
        - Notas
      summary: Actualizar un webhook
      description: Actualiza la configuración de un webhook existente.
      operationId: updateWebhook
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  type: string
                  description: ID del webhook a actualizar
                url:
                  type: string
                  format: uri
                  maxLength: 2048
                secret:
                  type: string
                  maxLength: 256
                  nullable: true
                events:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  maxItems: 20
                active:
                  type: boolean
                description:
                  type: string
                  maxLength: 500
                  nullable: true
      responses:
        '200':
          description: Webhook actualizado exitosamente
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Notas
      summary: Eliminar un webhook
      description: Elimina un webhook existente.
      operationId: deleteWebhook
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook eliminado exitosamente
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: Cookie de sesión de NextAuth.js
  schemas:
...
    Webhook:
      type: object
      required:
        - id
        - url
        - events
        - active
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: ID único del webhook
        url:
          type: string
          format: uri
          maxLength: 2048
          description: URL del webhook
        events:
          type: array
          items:
            type: string
          description: Eventos que escucha el webhook
        active:
          type: boolean
          description: Indica si el webhook está activo
        noteId:
          type: string
          nullable: true
          description: ID de nota específica (null para webhook global)
        description:
          type: string
          nullable: true
          maxLength: 500
          description: Descripción del webhook
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        totalDeliveries:
          type: integer
          description: Total de deliveries realizados
        recentStats:
          type: object
          properties:
            success:
              type: integer
            failed:
              type: integer
            total:
              type: integer
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Mensaje de error descriptivo
        code:
          type: string
          description: Código de error único
        details:
          type: object
          description: Detalles adicionales del error
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Parámetros de consulta inválidos'
            code: 'INVALID_QUERY_PARAMS'
    Unauthorized:
      description: No autenticado o sesión inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'No autorizado. Por favor, inicia sesión.'
            code: 'UNAUTHORIZED'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Recurso no encontrado'
            code: 'NOT_FOUND'
...
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Error interno del servidor'
            code: 'INTERNAL_SERVER_ERROR'